+++ src/Application/Query/Contact/ContactListView.php
@@ -1,0 +1,42 @@
+<?php
+
+namespace Proximum\Vimeet\Application\Query\Contact;
+
+class ContactListView
+{
+    /** @var array */
+    private $contactPreviewViews;
+
+    /** @var bool */
+    private $isItDDay;
+
+    /** @var bool */
+    private $isAccessControlEnabled;
+
+    public function __construct(bool $isAccessControlEnabled, bool $isItDDay, array $contactPreviewViews)
+    {
+        $this->isAccessControlEnabled = $isAccessControlEnabled;
+        $this->isItDDay = $isItDDay;
+        $this->contactPreviewViews = $contactPreviewViews;
+    }
+
+    public function getContactPreviewViews(): array
+    {
+        return $this->contactPreviewViews;
+    }
+
+    public function isItDDay(): bool
+    {
+        return $this->isItDDay;
+    }
+
+    public function isAccessControlEnabled(): bool
+    {
+        return $this->isAccessControlEnabled;
+    }
+
+    public function showCheckinStatus(): bool
+    {
+        return $this->isAccessControlEnabled() && $this->isItDDay();
+    }
+}